+++ docs/conf.py
@@ -18,12 +18,7 @@
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import datetime
-import os
-import sys
 
-sys.path.insert(0, os.path.abspath("../src/schemathesis"))
-
-# isort:imports-firstparty
 import schemathesis
 
 # -- General configuration ------------------------------------------------