+++ src/schemathesis/graphql.py
@@ -1,0 +1,5 @@
+# pylint: disable=unused-import
+# Public API
+from .specs.graphql import nodes
+from .specs.graphql.loaders import from_asgi, from_dict, from_file, from_path, from_url, from_wsgi
+from .specs.graphql.scalars import register_scalar